Moved oauth to service folder as this module acts like a service for other layers
