+++ backend/api/token.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter, Depends, HTTPException
 from backend.models.pydantic_models import Token
-from backend.oauth import authenticate_user, create_access_token
+from backend.service.oauth import authenticate_user, create_access_token
 from backend.config import ACCESS_TOKEN_EXPIRE_MINUTES
 from fastapi.security import OAuth2PasswordRequestForm
 from datetime import timedelta
