+++ backend/api/user.py
@@ -1,7 +1,7 @@
 from fastapi import APIRouter, Depends, HTTPException
 from backend.models.sqlalchemy_models import User
 from backend.db.sessions import session
-from backend.oauth import encrypt_password, get_current_user
+from backend.service.oauth import encrypt_password, get_current_user
 
 
 user_router = APIRouter()
